+++ packages/editor/src/views/viewport/pointer/ElementInFlowMoveDragHandler.ts
@@ -35,18 +35,13 @@ export class ElementInFlowMoveDragHandler implements DragHandler {
     const dst = findDropDestination(this.editorState, pickResult, [
       ...this.targets.keys(),
     ]);
-
-    if (dst) {
-      this.editorState.dropTargetPreviewRect = dst.parent.boundingBox;
-      this.editorState.dropIndexIndicator = dropDestinationIndicator(dst);
-    }
+    this.editorState.dropDestination = dst;
   }
 
   end(event: MouseEvent | DragEvent): void {
     this.editorState.snapper.clear();
     this.editorState.dragPreviewRects = [];
-    this.editorState.dropTargetPreviewRect = undefined;
-    this.editorState.dropIndexIndicator = undefined;
+    this.editorState.dropDestination = undefined;
 
     const dst = findDropDestination(
       this.editorState,
@@ -123,59 +118,4 @@ export function findDropDestination(
     parent,
     ref: inFlowChildren[index],
   };
-}
-
-function dropDestinationIndicator(
-  dst: DropDestination
-): [Vec2, Vec2] | undefined {
-  const { parent, ref } = dst;
-
-  const direction = parent.layoutDirection;
-  const inFlowChildren = parent.inFlowChildren;
-
-  let index = inFlowChildren.findIndex((o) => o === ref);
-  if (index < 0) {
-    index = inFlowChildren.length;
-  }
-
-  const parentRect = parent.boundingBox;
-  const parentPaddings = parent.computedPaddings;
-
-  if (direction === "x") {
-    let x: number;
-
-    if (index === 0) {
-      x = parentRect.left + parentPaddings.left;
-    } else if (index === inFlowChildren.length) {
-      const prev = inFlowChildren[index - 1];
-      x = prev.boundingBox.right;
-    } else {
-      const prev = inFlowChildren[index - 1];
-      const next = inFlowChildren[index];
-      x = (prev.boundingBox.right + next.boundingBox.left) / 2;
-    }
-
-    const y1 = parentRect.top + parentPaddings.top;
-    const y2 = parentRect.bottom - parentPaddings.bottom;
-
-    return [new Vec2(x, y1), new Vec2(x, y2)];
-  } else {
-    let y: number;
-
-    if (index === 0) {
-      y = parentRect.top + parentPaddings.top;
-    } else if (index === inFlowChildren.length) {
-      const prev = inFlowChildren[index - 1];
-      y = prev.boundingBox.bottom;
-    } else {
-      const prev = inFlowChildren[index - 1];
-      const next = inFlowChildren[index];
-      y = (prev.boundingBox.bottom + next.boundingBox.top) / 2;
-    }
-
-    const x1 = parentRect.left + parentPaddings.left;
-    const x2 = parentRect.right - parentPaddings.right;
-
-    return [new Vec2(x1, y), new Vec2(x2, y)];
-  }
 }
